+++ services/mqtt/config.go
@@ -29,8 +29,15 @@ type Config struct {
 	Username string `toml:"username" override:"username"`
 	Password string `toml:"password" override:"password,redact"`
 
-	// NewClientF is a function that returns a client for a given config.
-	NewClientF func(c Config) (Client, error) `toml:"-" override:"-"`
+	// newClientF is a function that returns a client for a given config.
+	// It is used exclusively for testing.
+	newClientF func(c Config) (Client, error) `override:"-"`
+}
+
+// SetNewClientF sets the newClientF on a Config.
+// It is used exclusively for testing.
+func (c *Config) SetNewClientF(fn func(c Config) (Client, error)) {
+	c.newClientF = fn
 }
 
 func NewConfig() Config {
@@ -52,8 +59,8 @@ func (c Config) Validate() error {
 // NewClient creates a new client based off this configuration.
 func (c Config) NewClient() (Client, error) {
 	newC := newClient
-	if c.NewClientF != nil {
-		newC = c.NewClientF
+	if c.newClientF != nil {
+		newC = c.newClientF
 	}
 	return newC(c)
 }