+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bugfixes
 
 - [#1250](https://github.com/influxdata/kapacitor/issues/1250): Fix VictorOps "data" field being a string instead of actual JSON.
+- [#1697](https://github.com/influxdata/kapacitor/issues/1697): Fix panic with MQTT toml configuration generation.
 
 ## v1.4.0-rc1 [2017-11-09]